% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_model.R
\name{fit_linear_model}
\alias{fit_linear_model}
\title{Fit a SVDE generalizef linear model}
\usage{
fit_linear_model(
  input_matrix,
  model_matrix,
  ncounts,
  inference_method = "SVI",
  method_specific_args = list(steps = 350L, lr = 0.1, gamma_lr = 0.1, cuda = TRUE,
    jit_compile = FALSE, full_cov = TRUE, batch_size = 10240L, prior_loc = 0.1,
    theta_bounds = c(1e-06, 1e+06), init_loc = 0.1, init_theta = 1000)
)
}
\arguments{
\item{input_matrix}{}

\item{model_matrix}{}

\item{ncounts}{}

\item{inference_method}{}

\item{method_specific_args}{}
}
\description{
Fit a SVDE generalizef linear model
}
