% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_test.R
\name{perform_test}
\alias{perform_test}
\title{Title}
\usage{
perform_test(
  fit_object,
  input_matrix,
  model_matrix,
  group1,
  group2,
  pval = 0.05,
  LFC = 0.5,
  pct = 0.3,
  filter = TRUE
)
}
\arguments{
\item{fit_object}{}

\item{input_matrix}{}

\item{model_matrix}{}

\item{group1}{}

\item{group2}{}

\item{pval}{}

\item{LFC}{}

\item{pct}{}

\item{filter}{}
}
\description{
Title
}
